requirements.txt
django==4.2.1
celery==5.2.7
python-dotenv==1.0.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
django-ses==3.5.0
psycopg2-binary==2.9.6
whitenoise==6.4.0
sentry-sdk==1.24.0
wagtail==5.0.1
django_extensions==3.2.1
django_storages==1.13.2

drf-rw-serializers==1.0.5
django-crum==0.7.9
Redis==4.6.0
# remove it
django-phonenumber-field==7.1.0
arrow==1.2.3
phonenumbers==8.13.13
Pillow==10.2.0
